nubus 8•24 GC: split the GC QuickDraw engine into its own file
display_card_824gc.c had grown to 3,750 lines spanning two very
different kinds of code.  Split it along the hardware boundary:

- display_card_824gc.c (1,413 lines) — the card shell: decl ROM,
  JMFB-family display registers, NuBus address dispatch, CB bring-up
  state machine, RPC transport, vtable/factory, object model.
- display_card_824gc_qd.c (2,175 lines) — "GC QuickDraw", the drawing
  engine that ran ON the Am29000 on real hardware: the DrawMultiObject
  interpreter, pixel cores, patterns + PixPat cache, conic and polygon
  rasterizers, region clip masks, the func $15 blit engine, and the
  font/text machinery.
- display_card_824gc_priv.h — internal header: the private card state,
  DRAM accessors, and the seven gc824_* entry points the shell calls
  (interp, stretchbits, cache_pixpat, pixpats_flush, font_download,
  font_caches_flush, reset_draw_state).

Pure code motion plus the entry-point renames; no behavior change.
Also fixes the two long-standing -Wtype-limits warnings in the
super-slot dispatch (offset-subtraction range test instead of a >= 0
comparison against a zero offset).

All 9 iicx-824gc* integration tests pass, including the pixel-exact
accel-vs-ROM oracle pairs; native and wasm builds are warning-clean.
